Revert "Add OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1817c4b2-5c34-46f4-85b4-05f059c799b7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia95bbfd95694eb1328e01779b37681ecc54df6a3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com