Revert "Add OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0e192b5d-d896-409d-86e4-91ca27575a87). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2d1cbfe9597225bc44f4eb1f58a97620c09a3d5c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com