Revert "Add OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/567f7012-e043-4b0f-ae1a-c2fd940b67c3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6095dccabbea89bf0b13becfb2d7c2abd20e10c8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com