Revert "Add OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f17f129-8e87-4dd8-a5e7-69bd8349068e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia9ac773ecb7be34f53f73c5ff9fb3d1e9b06f5a5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com