Revert "Add OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a8b2a8f5-945c-46d1-8b4b-9020950565e9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I12b1e2655811b1642ec65d0ee7d90b3b05e84ee5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com