Revert "Add OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/332adb8b-6cf6-4693-813f-8a825fc2d5d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0db846612b858ea54f6e3717f5e1a84b8265c250
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com