Revert "Add OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed67fcbe-bb6e-4a9d-9f54-9e0bc4ba54e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I63105a60903f748b7f5adfe0907d9d21342771d6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com