Revert "Add OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/071b8970-9709-4d0c-bef8-8c5d60236e24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia6f743775a6bc86ccba683d10b70d920967a5f60
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com