Revert "Add OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/945fc41c-ef52-4280-97b8-715225144b2c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaa94495a557557be7c1ff8f23ed240a7a1a199d9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com