Revert "Add OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9ad05ee0-daeb-451b-a577-47b4a75c6e9e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia70b68f60b033c64ac10b43d5bdb74a6dd239b33
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com