Revert "Add OWNERS file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd393e28-20af-4ca8-acf6-85de8c0bbdd1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I473284e9c5b1422aab0ab7a8e964dcd9a87d4f23
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com