Revert "PREUPLOAD.cfg to run crate tool."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e2b21796-1fb8-423a-bd91-947d14a800d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib6ca6d0f968f415c440ebe5c3e8576c65eac3aec
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com