Revert "crates directory with METADATA file."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e2b21796-1fb8-423a-bd91-947d14a800d8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia2c52f622d9f634dd04399cf9b44f98973fff881
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com