Revert "Upgrade itertools to the latest version."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9ba4afef-7ed9-43fc-8ad0-5e76c5ff5a2d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I63a0f4e3d677b9dcfffe104a1f9343e4b1913749
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com