Revert "Upgrade itertools to the latest version."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7afb1e64-145a-48a2-a7e7-572adf64980c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3b1c595c2f53a510e2dce04b82fc961f2230ce8a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com