Revert "Upgrade itertools to the latest version."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/12973fd9-94b3-42d5-a2df-eb163309b766). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1d9705c11ffa948643929acbab43bc4228215205
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com