Revert "Migrate 25 crates to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/62f79cc2-da9a-496e-8a3b-1b4cca8ed38b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id60019d64b12a9997aff98b50efbb7e79bb5f3b3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com