Revert "Upgrade tempfile to 3.12.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/750c7a0b-fe0d-45f0-95fa-7a6ce6b5ad3c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 366171449
Change-Id: Ibb6e77f173a5797b664c1d7d24ad72b9c92ba8d6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com