Revert "Upgrade tempfile to 3.12.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5c0c89aa-cbd7-4db3-8483-182d317307ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 366331210
Bug: 366332233
Change-Id: I14641bcdbb5f1d2933641e33e4030ed04586db4f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com