Revert "Migrate zeroize* to crate monorepo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4050339a-20a4-4cda-a9f8-5d006878ee0a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3f3431517f97f3987eb84f23022d1e92b14b0cf7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com