Revert "Migrate zeroize* to crate monorepo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb74f86f-5f3d-4b9d-bd50-bce5e4fd0a50). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifaa500f8b5a4a986b16d076f044561791997f2d5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com