Revert "Migrate zeroize* to crate monorepo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0f1c8904-c9c5-4623-917b-bab43694dca5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I501f9d1a9606ee2e154c2bc96aec65891ccf783d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com