Revert "Migrate 8 crates to monorepo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b69e12f3-e29d-4290-8aa2-f2b68f5b769b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373654024
Change-Id: I64c0d710a507223d3a12a47e78707b803e2f5c71
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com