Revert "Migrate 8 crates to monorepo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c754399d-91f9-41b3-bd4a-11e81cc3ea09). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373653717
Change-Id: Id8a079b2fd97d959f022e57518cbc1688ae7bd44
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com