Revert "Migrate bytemuck to monorepo, with upgrade to 1.16.3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6673782a-ff0c-4472-8160-af084c640a11). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 373725434
Change-Id: I26352e5c12962ecfd7a7f552bb9d13bba9b75f8d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com