Revert "Regenerate zeroize and zeroize_derive."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a5c0a381-a7fc-4318-912c-846003fed974). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia68923eb3ef1d241d425d59f4b85ccfc7d11dfe1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com