Revert "Migrate heck to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7bf18b54-6a8c-4fa7-88e8-13ccd61070e8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375231257
Change-Id: Ie6694edcaf9d77b48b057e70205c9f03dd24fc9b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com