Revert "Migrate heck to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8eb5980-aba6-4f68-a839-41d0abdef7ac). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375231378
Change-Id: I0fb5cce0834fe3723d725cf45a6cd893a343e0ad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com