Revert "Migrate heck to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6ce13a4c-b1a4-4cc0-ad8c-c097655dcab7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375231059
Change-Id: I65ed61d2cf2c76e5f14756608338c2e8be16257c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com