Revert "Migrate heck to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c2c49b8b-6f6a-4f02-ac65-b7bff2cdbad9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375231073
Change-Id: I221bb6bcfabb50e47d015bcde3f3d5d8592a9cd2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com