Revert "Migrate heck to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/408a3d79-8f46-49ef-9e2c-8b023962cc7b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375231738
Change-Id: I13dc8f637a1d8441b77dd8ecabd75b1d6b627a91
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com