Revert "Migrate heck to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aaf333dd-efc2-4c74-977f-0db01cb98719). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375231252
Change-Id: I65cbbd4b2d00f066a178ffc06e08ded7834199d1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com