Revert "Migrate heck to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b855dab5-af85-459f-8e76-57a821dda030). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375231394
Change-Id: I9c1f23531345a0a18590a103a16ab990b5fdcd77
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com