Revert "Migrate heck to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c55c7bf-7272-495d-8273-19300d2068b7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375232023
Change-Id: I66d9a7e3e99c202ab383e42e8ae061bd5238d295
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com