Revert "Migrate heck to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/43227c47-ae82-4fd2-af44-e244c8dd2301). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375231298
Change-Id: I5fdf496c1906337245c8498f3f7ab82332217851
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com