Revert "Migrate heck to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0f6fcfac-03ab-4e17-9023-6ace122f6636). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375231442
Change-Id: Ib064be84491824ab317e7693810ccb68df795157
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com