Revert "Migrate heck to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/729397fa-9d3f-4419-bcd6-27a2c4224875). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 375292143
Bug: 375299971
Change-Id: I262857cf639ef8657492eee62975b6fe63b3e7bf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com