Revert "Migrate zerocopy to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/268722a3-0721-4222-a32e-cada1a2f17bf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 377559880
Change-Id: Iab11f92d3462a84a65c9d6cbdbe66ec3738cbac4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com