Revert "Migrate zerocopy to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ab427ad9-188d-4a29-a2b2-fdd2ce58334d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 377559737
Change-Id: Ic5f2fcff79c733fbecfc27adb5a2a25c4af31243
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com