Revert "Migrate zerocopy to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1fac9dda-4591-4837-a44b-cdeed10f924f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 377579196
Change-Id: Id3400f995624d7444a8a4a2fdd91af1de1638b7b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com