Revert "Fix rules.mk for zeroize"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/af4f2839-6e09-4392-ac83-1fd66f037fdd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If1eb0eb7290aac47f5a01687c3b6b7d8883fa651
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com