Revert "Generate rules.mk for tinyvec via cargo_embargo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/af4f2839-6e09-4392-ac83-1fd66f037fdd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9819d1bd61857bd6d0aa54dc7b491906d4cea58d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com