Revert "Set min_sdk_version=29 for foreign-types and foreign-typ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/af4f2839-6e09-4392-ac83-1fd66f037fdd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3228b2b2008ef30de3b2914c6078881b2e0d076a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com