Revert "bytemuck crate: Add no_std variant"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9e8b1a36-8144-49ca-9360-e7dca09601a7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I83c8e9a8eadb5d1ae5d87c13e9c8d97ef294b1e1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com