Revert "Limit std patches to android_dylib only"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/49e01b79-3c72-4091-a395-b3bd498f9547). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1fe45a3721a2e3d6daa98d7f3aed430a25d78ea5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com