Revert "Limit std patches to android_dylib only"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81b6a602-633d-433e-8de0-58721b8da4f5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9c26cabe06008cead6a82855762872cee18bf963
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com