Revert "Migrate 9 newly imported crates to monorepo"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a85508b6-bfdc-428a-a349-2717ba99a371). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I466bbe735b5b676f91dddca8663700b622ea014d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com