Revert "Fix cargo_embargo.json and crate-list.txt for vm-memory ..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a85508b6-bfdc-428a-a349-2717ba99a371). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I353469a7604837d4d7d39b186916da86fa0d1d5c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com