Revert "Update syn to the latest version."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/13a26687-8a6e-4d95-9926-9f8bd74ae6e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I90f8760a8af9ae4475ce80b8e24a8d396207165e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com