Revert "Update syn to the latest version."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/13a26687-8a6e-4d95-9926-9f8bd74ae6e2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I29bbf2ff8e7203e947b90346557a0af70dd0aac8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com