Revert "Update quote to 1.0.37"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb7fd529-c7d1-48b8-94fb-22ba8b187711). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0db212fe299ebb03a2016e000f35d8889f3f6387
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com