Revert "Update libbpf-rs to 0.24.8"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/57c32b52-9a94-4e58-a47c-758e6babe310). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifb0b76b1ffbb6d02a469036d74a6da84bc55ca5c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com