Revert "Update bitreader to 0.3.11"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/57c32b52-9a94-4e58-a47c-758e6babe310). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5955513d20c99818b07636110f926dee1b03f52a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com