Revert "Update gdbstub to 0.7.3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6b88b5bf-9486-4c6a-ab80-a424367c58cb). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib9d9c0fc349c16fe46b868bfaaf3b128bc972acd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com