Revert "arbitrary: Disable tests to allow upgrade"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27538e04-ce8a-4c51-aaed-f2895c01c9ef). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388438703
Change-Id: I14b3a94ef05c60d54ac84777caac0ff635aefa7d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com