Revert "Regenerate all crates with .android-checksum.json alphab..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/27538e04-ce8a-4c51-aaed-f2895c01c9ef). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 388438703
Change-Id: I317c18dbdfffa19b60a074780d7ad42c4cdebde5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com