Revert "Add `min_sdk_version` property to criterion lib"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/69d4fc37-0eaa-4722-80c8-dfd0bb4f9716). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I77466d0b2b2b23bce3e7796e95188237aed4ea69
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com