Revert "Update memoffset to 0.9.1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/724d632c-467e-4b16-ae73-7122cd3bca03). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaab0a34225910b939b0d71a876b34e314e4558d5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com