Revert "Update memoffset to 0.9.1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9166b1db-467c-4e92-ad7a-f70848b18df4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic84cd0485ba433af6c919586674e55a82e27f99f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com