Revert "Update libz-sys to 1.1.20"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9166b1db-467c-4e92-ad7a-f70848b18df4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifa6eaffb0ef13253a2564858ad9a3cceab009088
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com