Revert "Update semver to 1.0.25"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c1159b4d-b1d6-4490-acf5-511e15c808df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id7140b39c18020374ecc417fee3e3c6687b64654
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com