Revert "Import is-terminal 0.4.13"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9ff0f077-f43b-43af-ad2b-58237b811d2e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5db19d109a1eed6889af2b40fceb3aac4661d394
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com