Revert "Import is-terminal 0.4.13"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/71668014-a3aa-4a12-a80f-ef18b6d651f2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2520fab2c2d7af9d8a38da5ac6fe86266e0b46e0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com