Revert "Import textdistance 1.1.1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8e16049-0abd-40d3-81ec-65c2deaa57b4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifabd6c951a25844c326cab33adaf17dc0ebb87ea
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com