Revert "Update async-trait to 0.1.85"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/470ed953-f92e-4be3-9409-4d504357b107). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2e89ec133282ca86fac2fbf5c67f78d0f8f2b1f3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com