Revert "Import textdistance 1.1.1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/36a3c43b-8635-49e5-a72c-6d370dd64dc1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ibde7c72c3eba5ad496ec930b66c0a7bcb515fbd9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com