Revert "Import is-terminal 0.4.13"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ecae5211-dcf3-4e06-af94-eb8d13e91088). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I77d5409994120ee11e71255969d512cb10da993f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com