Revert "Import textdistance 1.1.1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ecae5211-dcf3-4e06-af94-eb8d13e91088). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I454301dac5ee7ab4af47ce526098abbfc2a1b726
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com