Revert "Import 'diplomat_core' crate"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7446d715-40e6-468b-80ed-4f094d92be56). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ed5b0cdf2594c21e6237b89a67dec863fa23c60
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com