Revert "Fix license file for ident_case"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/00addb37-af74-4555-bb91-ab97871fcd67). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8fc954b81d93aca8a70a046abc0ec59c243402d5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com