Revert "Fix license files for virtio-queue."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/00addb37-af74-4555-bb91-ab97871fcd67). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I086a9c329f601e18bbc5e1b4b3faba85dec5a5c7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com