Revert "Update ryu to 1.0.19"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/00addb37-af74-4555-bb91-ab97871fcd67). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I771ad8b912410e1131f10da50ffe2f530acb0864
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com