Revert "Migrate libsqlite3-sys to monorepo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cdc3b379-871d-426d-9023-551e1c32ab24). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I677bbea07bed9449134d69088ae8baa9a25dbec1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com