Revert "'Import 'stable-deref-trait' crate"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/278cf864-0a7c-4066-aaf4-4723184ebb79). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I07bc3e615ec425592f5b737daef84c1f3197732c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com