Revert "'Import 'stable-deref-trait' crate"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9319a3d-a393-4ccb-8c28-848e18118115). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5b2854db6f845c64955a65b01b3d7b56e2fe3d53
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com