Revert "'Import 'stable-deref-trait' crate"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0f61b7b4-e1f4-4fc3-8447-d3e9ead1ea11). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia55bbf77b6aec87461b9ec95e3e54cd09baa1b8b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com