Revert "'Import 'stable-deref-trait' crate"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6dbd1d65-acaa-42f9-827d-d7b891252684). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4ec374cc276265d003b563615aee7599970495b4
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com