Revert "Add zerocopy 0.8 rules.mk, use Cargo to detect version cfg's"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c929927f-b18f-425a-8777-56a027e4ad7e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic20c1fa624d3f6aef10934947d51e955551e6137
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com