cov-mark: Instrument only in the _enable variant
cov-mark is used in our code base so unit tests can check that a
certain code path marked by `cov_mark::hit!` is actually called
(see docs in https://docs.rs/cov-mark/latest/cov_mark/).
The macro expands to atomic counters with the "enable" feature, and nothing when not enabled.
I'd like to enable it in tests and disable it for production builds,
hence providing two variants in the crate.

Bug: 377251092
Test: atest cov-mark_enable_test_tests_smoke  cov-mark_test_tests_smoke --host
Change-Id: I02349b5ccb61874732faf7ba8568d19200fca935