Revert "Import 'diplomat-runtime' crate"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/16c2d30b-8a30-4e57-9620-8ccbc48db80d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1595f7b79a0a553a266e39801e3c3eca3fda191f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com