Revert "Add missing rules.mk files."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d9f3cc5e-8a01-44d3-b16e-1c196fca6305). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If5ce5b2e6fe6488f38bc037e95ee7fb4b531cf57
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com