Revert "Update mls-rs-core to 0.18.1"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9b9005a2-4d63-4936-a870-e0c93276c82c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I17234d10b75113fdad99dae28b363b7d618c718c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com