Revert "Add missing rules.mk files."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/df9ecacc-7a32-49c1-860a-c534758bfbe5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I095f32711053849b041626afd2e6946c986c926a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com