Revert "Update once_cell to 1.20.3"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f525280e-3da8-4e79-9c39-7b21384e4d10). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1642bd8a2448819dcc948337ab9294ed594942fd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com