Revert "Update all METADATA files."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e699716f-8dc2-4bfc-bea1-b5f010b7b1d0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e8037990d3a143cfda1cf0b1fcc0980fc36af67
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com