Revert "Migrate critical-section to android-crates-io"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/61c43af9-b53d-4058-b037-9b74ce0c34ed). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a3304392ca44c7753ff87b5cc8170290ed34f0a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com