Revert "cov-mark: Add a variant without the "enable" feature"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e54d3fc3-5147-49bf-828e-1b89f66a09ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I21dcf37899777d81d35ab112c2792c6a9ef93e98
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com