Revert "Update serde_json to 1.0.139"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bfee2215-ae30-44c2-a654-f770cc83d293). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I902d675f74b186cc3da5cda689ecf49edfbec657
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com