Revert "Update etherparse to 0.17.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cb342be8-abff-42bb-b878-bfb100882d6a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I414ba74a6b7e451294110c13ddcfea822572ac0c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com