Revert "Update rusqlite to 0.32.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8d6dd7d-e64a-4977-92e2-68f7a8581a7f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I105e8c98c4537be9ca4c6fca0255f5aa53145166
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com