Revert "Copy hashbrown 0.12.3 to extra_versions."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d8d6dd7d-e64a-4977-92e2-68f7a8581a7f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4994390873617ce6db9ca749bc31b34bfb78ff2b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com