Revert "Regenerate userfaultfd-sys and vulkano."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c9905374-c450-4000-aa9d-0f3890383192). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icda6ff04a1849de9c431ce570543fb1b6aee20ab
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com