Revert "Update tinystr"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9cf3a7f4-0a55-487f-999a-437da8092dfa). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iebb3223bdea5bd84123ec01d39ba7ca23460ed6a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com