Revert "Update gdbstub to 0.7.5"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/97968232-0ed4-41db-b8f5-b6681db28105). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ifd5cb264d2e3ee933e6c8113a42a95b52e4a3416
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com