Revert "Update async-trait to 0.1.87"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/502ee89d-e85b-4539-b2b4-ff5f26cd633b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7881cf211ca9ab4a69df358ae74f65f3d0d08295
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com