Revert "Update serde to 1.0.218"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/502ee89d-e85b-4539-b2b4-ff5f26cd633b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia16e232c254dbf58c0dc3a25714e3da7ecde1391
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com