Revert "Update taffy to 0.7.6"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7f84c205-355b-48b0-9ef1-a7c7e05b2e6c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I24c0f120edbec48274a7e833e64c4c8bb07fb725
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com