Revert "Update serde_json to 1.0.140"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/6fbf8846-07a0-41bb-b5f2-20a21f24587f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4204e68ed9447d9d8247bfa1a59982bfd54aca72
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com