Revert "Set min_sdk_version for 8 crates"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8fa97018-30ed-410f-b4f6-20a780b00ec0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icef492264ec7d5cfe6bfb1529feea1fd4455bc83
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com