Revert "Update quote to 1.0.39"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3e5a1999-759f-4f1a-b6f9-f9f55f411934). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib1f6cf1dbbd7e3c769dc430faffb6df0c2cd82cf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com