Revert "Update android_logger to 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca996cbe-6670-49db-a55f-c8a50569216d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I815a07e4f2c1bb37098bb592f8ddb39cc804149a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com