Revert "android_logger: patch a crash in quiche"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca996cbe-6670-49db-a55f-c8a50569216d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id44f3d5e4282a209e8379437bc068beadef9ca18
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com