Revert "log: rerun crate_tool after patch update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca996cbe-6670-49db-a55f-c8a50569216d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I747acb3a5c93cb0105e507ebc5078075e3766739
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com