Revert "Update android_log_sys to 0.3.2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca996cbe-6670-49db-a55f-c8a50569216d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I04d8ff50167d107b0b7766da806668f86365de3f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com