Revert "env_filter: lower min_sdk_version to 29"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca996cbe-6670-49db-a55f-c8a50569216d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I73e87df58a06ec4f88086083fb4c3fbac4f19f4f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com