Revert "log: update aosp/2717613 patch for version update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ca996cbe-6670-49db-a55f-c8a50569216d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I81ef37534c61a0b3a297dec4c194b472ab830cb2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com