Revert "log: update aosp/2717613 patch for version update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ecbc8f7-4b23-4016-9371-bdf9b55e1a05). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id9b7d89801a3024e1ae3f5adcb7d8d4fe9d0234e
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com