Revert "Update android_log_sys to 0.3.2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ecbc8f7-4b23-4016-9371-bdf9b55e1a05). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icbdc57979c28b09c2ac5c0040c61bf4115b58348
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com