Revert "android_logger: patch a crash in quiche"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ecbc8f7-4b23-4016-9371-bdf9b55e1a05). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4de95e2d0d1ec1733ef3e419e1b8ca57830c7e5f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com