Revert "Update android_logger to 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1ecbc8f7-4b23-4016-9371-bdf9b55e1a05). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icf666c2bea294b5432cdbfcab171ef1e6dd29f52
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com