Revert "log: update aosp/2717613 patch for version update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a2e81d6-7537-4d98-88d9-74a727ac96c5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3658efadf0b3fb0afcb7f504f720e6bed6c6c4ed
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com