Revert "log: rerun crate_tool after patch update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a2e81d6-7537-4d98-88d9-74a727ac96c5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iaf3cf6f7af18c48e3a4730b25c947b035f5d5283
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com