Revert "env_filter: lower min_sdk_version to 29"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8a2e81d6-7537-4d98-88d9-74a727ac96c5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7dd4407d1dc403efcddeb64b1e5c057008460037
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com