Revert "android_logger: patch a crash in quiche"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb885655-6c4b-4016-aa7a-a82e65c129e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0191523f873ca7680c03abce1a3c6b7733ec6f0d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com