Revert "Update android_log_sys to 0.3.2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb885655-6c4b-4016-aa7a-a82e65c129e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie6899825710ba3993d5d3e41be94d3cdfe607068
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com