Revert "log: rerun crate_tool after patch update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb885655-6c4b-4016-aa7a-a82e65c129e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I22c04700f0ec06a3496191f0ef9863da16a9e433
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com