Revert "android_logger: update aosp/2717316 patch for 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb885655-6c4b-4016-aa7a-a82e65c129e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5229bce66f718104d7f91748a471b50157a356aa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com