Revert "env_filter: lower min_sdk_version to 29"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb885655-6c4b-4016-aa7a-a82e65c129e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iace493bae2a14f436d3059bab4ccf98f27afd0ff
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com