Revert "Update android_logger to 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/eb885655-6c4b-4016-aa7a-a82e65c129e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iab17dc61e58487c7cb4bf0cf021a9ce7516f4022
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com