Revert "Update android_logger to 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/744651f7-f54e-4c4f-bbf6-3cd89e9c66d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e5023ea201e80fc18888dd28bc299c3e997b46b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com