Revert "env_filter: lower min_sdk_version to 29"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/744651f7-f54e-4c4f-bbf6-3cd89e9c66d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9fc1ab9c3b64ace28297b333ec76db4427fee38a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com