Revert "log: rerun crate_tool after patch update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/744651f7-f54e-4c4f-bbf6-3cd89e9c66d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id558f958c9c98dd2629f0a95d48bc4e3d2475f55
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com