Revert "Update android_log_sys to 0.3.2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/744651f7-f54e-4c4f-bbf6-3cd89e9c66d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icc59a1eae94f64d8a043d79b34df9eaeed0a5db6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com