Revert "android_logger: update aosp/2717316 patch for 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/744651f7-f54e-4c4f-bbf6-3cd89e9c66d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I191ffc7829ed4424015c3e0a39008bda5fa2ead6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com