Revert "android_logger: update aosp/2717316 patch for 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/be5f2b7d-8a65-468d-97fc-0b8d8b5e7797). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id218a385e280422aa5842f0116fe3954ff2cb5b2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com