Revert "android_logger: patch a crash in quiche"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/be5f2b7d-8a65-468d-97fc-0b8d8b5e7797). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2ca93029e60fee701dff34de825d8eb02c9e734d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com