Revert "Update android_logger to 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/be5f2b7d-8a65-468d-97fc-0b8d8b5e7797). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0db7b34726cf62c949f6daf8eb34bd6a95771393
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com