Revert "log: update aosp/2717613 patch for version update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/be5f2b7d-8a65-468d-97fc-0b8d8b5e7797). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I371abb4f1412cc91030072c4803a438db689ea3d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com