Revert "Update android_log_sys to 0.3.2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/be5f2b7d-8a65-468d-97fc-0b8d8b5e7797). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5adb4cc27b061756db746a9e1a2a9dd0dc3f1d24
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com