Revert "log: rerun crate_tool after patch update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/be5f2b7d-8a65-468d-97fc-0b8d8b5e7797). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3c64370fa0d1b5786352e6f354ff2ea298c82e1d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com