Revert "Update safe-mmio to 0.2.1."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a078e371-3859-4417-a8d9-1e349396609b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I63acf404d15af62f152da54124b48892e959f49d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com