Revert "Update android_logger to 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/24cf258c-4cd5-4492-9529-e6e21fb5b381). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I561116c6ae271d2a8bfa115fc8b3f74a4f3ef716
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com