Revert "env_filter: lower min_sdk_version to 29"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/24cf258c-4cd5-4492-9529-e6e21fb5b381). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8374e9408f246d3a77f0e276bb995f11bb162a86
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com