Revert "Update android_log_sys to 0.3.2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/24cf258c-4cd5-4492-9529-e6e21fb5b381). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1123ba748134346bc4f75a3a5d19baec4392d8b5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com