Revert "android_logger: patch a crash in quiche"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/24cf258c-4cd5-4492-9529-e6e21fb5b381). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I73c901ae5dc05ac82393550cc161077e293e2734
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com