Revert "log: rerun crate_tool after patch update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/24cf258c-4cd5-4492-9529-e6e21fb5b381). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8d434fc0e15dcf442d696db1f463631db8ea8f6b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com