Revert "android_logger: update aosp/2717316 patch for 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/24cf258c-4cd5-4492-9529-e6e21fb5b381). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I58c60a02f3d5fa2a2e3df8e797db6b4f7c1aff61
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com