Revert "Update safe-mmio to 0.2.1."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1f75818d-9b7f-4710-bdd4-22ce48396688). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia2fe987f5990c51bfa4a8a6f56cdf806b43feb5b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com