Revert "Update safe-mmio to 0.2.1."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/25a667ae-eb23-41cd-8285-8d2b49a15000). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6ec4d0ce645153a1e0460865d63579af162f42dd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com