Revert "Update safe-mmio to 0.2.1."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5a058001-f824-4bd9-8a42-d1692bc359f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I62cf94135cf60c166dc3ff3266626795461dbd6f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com