Revert "Update safe-mmio to 0.2.1."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/07df6fb0-501b-4ce5-8577-59607dae606d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia7fb3f216a9efb558bcfb12c158ca866e73f35c1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com