Revert "android_logger: update aosp/2717316 patch for 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2234f055-e2a2-4f7e-b077-c9a8a00ab607). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic92890a3adf48542ad12938c78ebae95f614a1b2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com