Revert "Update android_logger to 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2234f055-e2a2-4f7e-b077-c9a8a00ab607). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia3b3aa5e0c39f6d0a6f4038e550152ddc7a2a999
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com