Revert "Update android_log_sys to 0.3.2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2234f055-e2a2-4f7e-b077-c9a8a00ab607). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I9d3c7a35683f806228d2677f792908685cd262f1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com