Revert "android_logger: patch a crash in quiche"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2234f055-e2a2-4f7e-b077-c9a8a00ab607). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I96cddc1f37c4110c442a9a829a10ffcc41d282cb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com