Revert "log: rerun crate_tool after patch update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2234f055-e2a2-4f7e-b077-c9a8a00ab607). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5ae24686ae786e21ec47bc9e911211048e17e7e7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com