Revert "Update safe-mmio to 0.2.1."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9a320c92-c5bb-4721-822c-589365972d3f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I614b3b26e98c1d565ff864c933c898803fedb408
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com