Revert "Import 'cargo-platform' crate"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2abd088a-c63a-4e12-b0c5-bd87b2930353). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8fe1feb36726e5c2ba94bbdb10986e69d21e887a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com