Revert "Import 'ipp' crate"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c32208fa-5ae5-45cd-9c1d-48d239dc4ad3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I13c7ab3a0985a2496ce2a9238882cfe8c06656b9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com