Revert "Update libz-sys to 1.1.22"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fd156c97-b0d5-4b95-9607-d7abc92c9a9e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I604dce4080ba4ac06632dd692182032d5a1187e9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com