Revert "log: rerun crate_tool after patch update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d4b0235-7911-44ce-87b0-b8f9cb99eda1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id318f51840b44a86d5bbf4834cb4a9483ebb38a1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com