Revert "android_logger: patch a crash in quiche"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d4b0235-7911-44ce-87b0-b8f9cb99eda1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iba6bf15a8c562de7c47d978baca931c52d2853eb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com