Revert "Update android_log_sys to 0.3.2"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d4b0235-7911-44ce-87b0-b8f9cb99eda1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia604cdbd2c2806df2ab3b1542e76fa06c17ee9fc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com