Revert "log: update aosp/2717613 patch for version update"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d4b0235-7911-44ce-87b0-b8f9cb99eda1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I48467c58c91c3bd3378bee779775954eef1905f1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com