Revert "android_logger: update aosp/2717316 patch for 0.15.0"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4d4b0235-7911-44ce-87b0-b8f9cb99eda1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3b0922650a40a60de1222dad7e1a71d6d2058133
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com